+++ paper/paper.bib
@@ -236,7 +236,7 @@ publisher: Nature Publishing Group},
 	rights = {© 2019 {WILEY}‐{VCH} Verlag {GmbH} \& Co. {KGaA}, Weinheim},
 	issn = {2513-0390},
 	url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/adts.201900015},
-	doi = {https://doi.org/10.1002/adts.201900015},
+	doi = {10.1002/adts.201900015},
 	pages = {1900015},
 	number = {7},
 	journaltitle = {Advanced Theory and Simulations},
