### 2023-09-13 ecc7d84
`paper.md` cleanup
