+++ paper/paper.md
@@ -25,7 +25,7 @@ affiliations:
    index: 2
  - name: The Faraday Institution, Didcot, United Kingdom
    index: 3
- - name: Imperial College London, London, United Kingdom
+ - name: Department of Materials, Imperial College London, London, United Kingdom
    index: 4
 date: 6 June 2023
 bibliography: paper.bib
@@ -35,36 +35,36 @@ bibliography: paper.bib
 
 The electronic band structure is an important property for
 understanding and designing solid crystalline materials in many fields
-such as photovoltaic, catalysis, thermoelectrics and transparent conducting
+such as photovoltaic, catalytic, thermoelectric and transparent-conducting
 materials. Obtaining the band structure for an ideal crystal through first-principles
-density functional theory calculation is a well-established routine operation.
-However, the materials of interest are often complex and the simulation cells may contain multiple  primitive
-cells of the archetypal structure - when modelling disordered materials or defective structures, for example.
-Repeating the unit cell in the real space result in folded band structures, as illustrated in \autoref{fig:figure1}, which complicates its interpretation and analysis.
+density functional theory (DFT) calculations is a well-established routine operation.
+However, the materials of interest are often complex and the simulation cells may contain multiple primitive
+cells of the archetypal structure when, for example, modelling disordered or defective materials.
+Repeating the unit cell in real space results in folded band structures, as illustrated in \autoref{fig:figure1}, complicating its interpretation and analysis.
 Band structure unfolding maps the electronic structure from supercell calculations back to the reciprocal lattice of the primitive cell,
-thereby enabling researchers to understand structure-property and compare the effect of various crystal imperfections on an equal footing.
-With `easyunfold`, we provide a simple, easy to use, yet powerful and flexible tool which helps from input file generation to publication-quality plotting for band structure unfolding using plane wave density functional theory codes.
+thereby enabling researchers to understand structure-property relationships and compare the effect of various crystal imperfections on an equal footing.
+With `easyunfold`, we provide a simple, easy-to-use, yet powerful and flexible tool which implements the band structure unfolding workflow using plane-wave DFT codes, from input file generation to publication-quality plotting.
 
-![Schematics of band folding in a 2D lattice: (a) The primitive cell in the real space; (b) The $2\times2$ supercell in the real space. (c) A slice of the band structure in primitive cell's Brillouin zone; (d) The same slice in the supercell. \label{fig:figure1}](figure1.png){width=100mm}
+![Schematics of band folding in a 2D lattice: (a) The primitive cell in real space; (b) The $2\times2$ supercell in real space. (c) A slice ($\varepsilon$) of the band structure in the Brillouin zone of the primitive cell; (d) The same band structure slice ($\varepsilon$) in the Brillouin zone of the supercell. Band structure unfolding is the process of transforming the band structure slice in (d) back to Brillouin zone of the primitive cell (i.e. the form shown in (c)). \label{fig:figure1}](figure1.png){width=100mm}
 
 # Statement of need
 
 There are existing packages that provide band structure unfolding capabilities, such as `BandUp`
 [@bandup], and `VaspBandUnfolding` [@vaspbandunfolding].
-`easyunfold` is written in Python with a focus of easy-to-use, data provenance, reproducibility, and generating publication-quality figures.
[…]
