### 2023-09-13 06d04e7
Update `paper.bib` (subscripts and VODPs ref)
