+++ paper/paper.bib
@@ -78,7 +78,7 @@ Vegge and Lasse Vilhelmsen and Michael Walter and Zhenhua Zeng and Karsten W Jac
 }
 
 @article{Popescu:2009,
-	title = {Extracting \${E}\$ versus \${\textbackslash}stackrel\{{P}{\textbackslash}vec\}\{k\}\$ effective band structure from supercell calculations on alloys and impurities},
+	title = {Extracting {$E$} versus {$\vec{k}$} effective band structure from supercell calculations on alloys and impurities},
 	volume = {85},
 	url = {https://link.aps.org/doi/10.1103/PhysRevB.85.085201},
 	doi = {10.1103/PhysRevB.85.085201},
@@ -112,11 +112,8 @@ Vegge and Lasse Vilhelmsen and Michael Walter and Zhenhua Zeng and Karsten W Jac
   url = {https://github.com/QijingZheng/VaspBandUnfolding}
 }
 
-
-
-
 @article{huang:2022,
-	title = {Strong absorption and ultrafast localisation in {NaBiS}2 nanocrystals with slow charge-carrier recombination},
+	title = {Strong absorption and ultrafast localisation in {NaBiS$_2$} nanocrystals with slow charge-carrier recombination},
 	volume = {13},
 	rights = {2022 The Author(s)},
 	issn = {2041-1723},
@@ -138,7 +135,7 @@ Publisher: Nature Publishing Group},
 }
 
 @article{nicolson:2023,
-	title = {Interplay of Static and Dynamic Disorder in the Mixed-Metal Chalcohalide Sn2SbS2I3},
+	title = {Interplay of Static and Dynamic Disorder in the Mixed-Metal Chalcohalide {Sn$_2$SbS$_2$I$_3$}},
 	issn = {0002-7863},
 	url = {https://doi.org/10.1021/jacs.2c13336},
 	doi = {10.1021/jacs.2c13336},
@@ -153,7 +150,7 @@ Publisher: Nature Publishing Group},
 }
 
 @article{wang:2022,
-	title = {Cation disorder engineering yields {AgBiS}2 nanocrystals with enhanced optical absorption for efficient ultrathin solar cells},
+	title = {Cation disorder engineering yields {AgBiS$_2$} nanocrystals with enhanced optical absorption for efficient ultrathin solar cells},
 	volume = {16},
 	rights = {2022 The Author(s), under exclusive licence to Springer Nature Limited},
 	issn = {1749-4893},
@@ -245,3 +242,18 @@ publisher: Nature Publishing Group},
 	date = {2019},
 
 }
+
+@article{kavanagh_frenkel_2022,
+  title = {Frenkel {{Excitons}} in {{Vacancy-Ordered Titanium Halide Perovskites}} ({Cs$_2$TiX$_6$})},
+  author = {Kavanagh, Se{\'a}n R. and Savory, Christopher N. and Liga, Shanti M. and Konstantatos, Gerasimos and Walsh, Aron and Scanlon, David O.},
+  year = {2022},
+  month = dec,
+  journal = {The Journal of Physical Chemistry Letters},
+  volume = {13},
+  number = {47},
+  pages = {10965--10975},
+  publisher = {{American Chemical Society}},
+  doi = {10.1021/acs.jpclett.2c02436},
+  urldate = {2023-01-03},
+}
+
