+++ paper/paper.bib
@@ -256,4 +256,3 @@ publisher: Nature Publishing Group},
   doi = {10.1021/acs.jpclett.2c02436},
   urldate = {2023-01-03},
 }
-
