### 2023-09-14 b6165a5
Fix a bug when parsing PROCAR with spin polarsation
The list of parsed kpoints should only be upated after parsing is
completed, otherwise the second set of spin down projections is
completely skipped.
