+++ easyunfold/procar.py
@@ -16,12 +16,13 @@ from easyunfold import __version__
 class Procar(MSONable):
     """Reader for PROCAR file"""
 
-    def __init__(self, fobjs_or_paths=None, is_soc=False):
+    def __init__(self, fobjs_or_paths=None, is_soc=False, normalise=False):
         """
         Read the PROCAR file from a handle or path
 
         :param fobjs_or_paths:  Either a string or list of file-like objs or paths
         :param is_soc: Whether the PROCAR is from a calculation with spin-orbit coupling
+        :param normalise: Whether normalise the projection for every band or not
         """
         self._is_soc = is_soc
         self.eigenvalues = None
@@ -36,6 +37,7 @@ class Procar(MSONable):
         self.proj_data = None
         self.header = None
         self.proj_xyz = None
+        self.normalise = normalise
 
         # Read the PROCAR
         if isinstance(fobjs_or_paths, (str, Path)):
@@ -56,19 +58,31 @@ class Procar(MSONable):
             raise ValueError(f'Mismatch in number of ions in PROCARs supplied: ({nion} vs {self.nion})!')
 
         # Count the number of data lines, these lines do not have any alphabets
-        proj_data, energies, kvecs, kweights, occs = [], [], [], [], []
+        proj_data, energies, kidx_internal, kvecs, kweights, occs = [], [], [], [], [], []
         tot_count = 0  # count the instances of lines starting with "tot" -> (4 + 1) * nbands * nkpts for SOC calcs
         fobj.seek(0)
 
         line = fobj.readline()
+        # Counter for the number of sections in the PROCAR
+        section_counter = 1
+        _last_kid = 0
+        this_procar_parsed_kpoints = set()
         while line:
             if line.startswith(' k-point'):
                 line = re.sub(r'(\d)-', r'\1 -', line)
                 tokens = line.strip().split()
+                _kid = int(tokens[1])
+                kidx_internal.append(_kid)
+
+                # Check if the internal k index counter decreases - if so we have entered a second section
+                if _kid < _last_kid:
+                    section_counter += 1
+                _last_kid = _kid
+
                 kvec = tuple(round(float(val), 5) for val in  # tuple to make it hashable
                              tokens[-6:-3])  # round to 5 decimal places to ensure proper kpoint matching
                 if kvec not in parsed_kpoints:
-                    parsed_kpoints.add(kvec)
+                    this_procar_parsed_kpoints.add(kvec)
                     kvecs.append(list(kvec))
                     kweights.append(float(tokens[-1]))
                 else:
@@ -107,21 +121,21 @@ class Procar(MSONable):
 
         proj_data = np.array(proj_data, dtype=float)
 
-        # redetermine nkpts in case some were skipped due to already being parsed
         nkpts = len(kvecs)
 
-        self.nspins = proj_data.shape[0] // (self.nion * nbands * nkpts)
-        self.nspins //= 4 if self._is_soc else 1
+        # When there are multiple spins, the data from the second spin are located after that of the first spin
+        # Hence, the number of spins is simply the number of sections
+        self.nspins = section_counter
 
         # Reshape
-        occs.resize((self.nspins, nkpts, nbands))
-        kvecs.resize((self.nspins, nkpts, 3))
-        kweights.resize((self.nspins, nkpts))
-        eigenvalues.resize((self.nspins, nkpts, nbands))
+        occs.resize((self.nspins, nkpts // self.nspins, nbands))
+        kvecs.resize((self.nspins, nkpts // self.nspins, 3))
+        kweights.resize((self.nspins, nkpts // self.nspins))
[…]
