### 2023-09-14 fe546df
PROCAR normalisation to be on by default
To make it more consistent
