### 2023-09-14 d7f8da6
Some small tidy up
