### 2023-09-14 6f2f7f8
Remove unused kidx_internal
