+++ easyunfold/procar.py
@@ -66,7 +66,7 @@ class Procar(MSONable):
         # Counter for the number of sections in the PROCAR
         section_counter = 1
         _last_kid = 0
-        this_procar_parsed_kpoints = set()
+        this_procar_parsed_kpoints = set()  # set with tuples of parsed (kvec tuple, section_counter) for this PROCAR
         while line:
             if line.startswith(' k-point'):
                 line = re.sub(r'(\d)-', r'\1 -', line)
@@ -78,10 +78,10 @@ class Procar(MSONable):
                     section_counter += 1
                 _last_kid = _kid
 
-                kvec = tuple(round(float(val), 5) for val in  # tuple to make it hashable
-                             tokens[-6:-3])  # round to 5 decimal places to ensure proper kpoint matching
-                if kvec not in parsed_kpoints:
-                    this_procar_parsed_kpoints.add(kvec)
+                kvec = tuple(round(float(val), 5) for val in tokens[-6:-3]  # tuple to make it hashable
+                            )  # round to 5 decimal places to ensure proper kpoint matching
+                if (kvec not in parsed_kpoints and (kvec, section_counter) not in this_procar_parsed_kpoints):
+                    this_procar_parsed_kpoints.add((kvec, section_counter))
                     kvecs.append(list(kvec))
                     kweights.append(float(tokens[-1]))
                 else:
@@ -90,7 +90,7 @@ class Procar(MSONable):
                         line = fobj.readline()
                     continue
 
-            elif not re.search(r'[a-zA-Z]', line) and line.strip() and len(line.strip().split()) - 2 == len(self.proj_names):
+            elif (not re.search(r'[a-zA-Z]', line) and line.strip() and len(line.strip().split()) - 2 == len(self.proj_names)):
                 # only parse data if line is expected length, in case of LORBIT >= 12
                 proj_data.append([float(token) for token in line.strip().split()[1:-1]])
 
@@ -134,7 +134,13 @@ class Procar(MSONable):
 
         # Reshape the array
         if self._is_soc is False:
-            proj_data = proj_data.reshape((self.nspins, nkpts // self.nspins, nbands, self.nion, len(self.proj_names)))
+            proj_data = proj_data.reshape((
+                self.nspins,
+                nkpts // self.nspins,
+                nbands,
+                self.nion,
+                len(self.proj_names),
+            ))
             proj_xyz = None
         else:
             proj_data = proj_data.reshape((self.nspins, nkpts, nbands, 4, self.nion, len(self.proj_names)))
@@ -151,9 +157,18 @@ class Procar(MSONable):
             proj_xyz /= proj_sum
 
         # Update the parsed kpoints
-        parsed_kpoints.update(this_procar_parsed_kpoints)
+        parsed_kpoints.update({kvec_section_counter_tuple[0] for kvec_section_counter_tuple in this_procar_parsed_kpoints})
 
-        return self.nspins, occs, kvecs, kweights, eigenvalues, proj_data, proj_xyz, parsed_kpoints
+        return (
+            self.nspins,
+            occs,
+            kvecs,
+            kweights,
+            eigenvalues,
+            proj_data,
+            proj_xyz,
+            parsed_kpoints,
+        )
 
     def normalise_projs(self, proj_data):
         """
@@ -201,8 +216,16 @@ class Procar(MSONable):
                 self._read_header_nion_proj_names(fobj)
 
             current_nspins = self.nspins  # check spin consistency between PROCARs
-            nspins, occs, kvecs, kweights, eigenvalues, proj_data, proj_xyz, parsed_kpoints = self._read(fobj,
-                                                                                                         parsed_kpoints=parsed_kpoints)
+            (
+                nspins,
+                occs,
[…]
