### 2023-09-14 2b9727a
Update tests (revert to previous numbers with fix, and explicitly test expected normalisation behaviour)
