### 2023-09-14 7bb1e13
Small fix for k-weighting (though rarely used) and add test
