### 2023-09-14 06f3c84
Use V_O^0 figure rather than Si in README
