+++ _sources/index.md
@@ -22,9 +22,9 @@ For the methodology of supercell band unfolding, see
 [here](https://link.aps.org/doi/10.1103/PhysRevB.85.085201).
 
 ### Example Outputs
-|    Cs₂(Sn/Ti)Br₆ Vacancy-Ordered Perovskite Alloys    |                Symmetry-broken Si Supercell                 |
-|:-----------------------------------------------------:|:-----------------------------------------------------------:|
-|   <img src="img/CSTB_easyunfold.gif" height="400"/>   | <img src="../examples/Si222/unfold_tall.png" height="400"/> |
+|    Cs₂(Sn/Ti)Br₆ Vacancy-Ordered Perovskite Alloys    |                        Oxygen Vacancy (*V*ₒ⁰) in MgO                         |
+|:-----------------------------------------------------:|:---------------------------------------------------------------------------:|
+|   <img src="img/CSTB_easyunfold.gif" height="400"/>   | <img src="../examples/MgO/unfold_project_MgO_v_O_0_tall.png" height="400"/> |
 
 ## Usage
 
