### 2023-10-11 #5841 @mbarzegary
Hi @lucydot @utf @srmnitc 👋
Considering that you have already published your software in JOSS, would any of you be willing to review this submission for [JOSS](https://joss.theoj.org/about)? As you know, we carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.05841/joss.05841/10.21105.joss.05841.pdf
Software: https://github.com/SMTG-UCL/easyunfold
