### 2023-10-17 #5841 @lucydot
@mbarzegary I am also editing for JOSS; I suggest that some of the authors from this package may be able to help review: https://joss.theoj.org/papers/10.21105/joss.04014
