### 2023-10-23 #5974 @editorialbot
[REVIEW]: easyunfold: A Python package for unfolding electronic band structures
**Submitting author:** @zhubonan (<a href="http://orcid.org/0000-0001-5601-6130">Bonan Zhu</a>)
**Repository:** https://github.com/SMTG-UCL/easyunfold
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.3.5
