### 2023-10-23 bab649f
Update paper reference
