+++ paper/paper.bib
@@ -256,3 +256,19 @@ publisher: Nature Publishing Group},
   doi = {10.1021/acs.jpclett.2c02436},
   urldate = {2023-01-03},
 }
+
+@article{liga_mixed-cation_2023,
+	title = {Mixed-{Cation} {Vacancy}-{Ordered} {Perovskites} ({Cs$_2$Ti$_{1–x}$Sn$_x$X$_6$}; {X} = {I} or {Br}): {Low}-{Temperature} {Miscibility}, {Additivity}, and {Tunable} {Stability}},
+	issn = {1932-7447},
+	shorttitle = {Mixed-{Cation} {Vacancy}-{Ordered} {Perovskites} ({Cs$_2$Ti$_{1–x}$Sn$_x$X$_6$}; {X} = {I} or {Br})},
+	url = {https://doi.org/10.1021/acs.jpcc.3c05204},
+	doi = {10.1021/acs.jpcc.3c05204},
+	abstract = {Lead toxicity and poor stability under operating conditions are major drawbacks that impede the widespread commercialization of metal–halide perovskite solar cells. Ti(IV) has been considered as an alternative species to replace Pb(II) because it is relatively nontoxic and abundant and its perovskite-like compounds have demonstrated promising performance when applied in solar cells (η {\textgreater} 3\%), photocatalysts, and nonlinear optical applications. Yet, Ti(IV) perovskites show instability in air, hindering their use. On the other hand, Sn(IV) has a similar cationic radius to Ti(IV), adopting the same vacancy-ordered double perovskite (VODP) structure and showing good stability in ambient conditions. We report here a combined experimental and computational study on mixed titanium–tin bromide and iodide VODPs, motivated by the hypothesis that these mixtures may show a stability higher than that of the pure titanium compositions. Thermodynamic analysis shows that the cations are highly miscible in these vacancy-ordered structures. Experimentally, we synthesized mixed titanium–tin VODPs as nanocrystals across the entire mixing range x (Cs2Ti1–xSnxX6; X = I or Br), using a colloidal synthetic approach. Analysis of the experimental and computed absorption spectra reveals weak hybridization and interactions between Sn and Ti octahedra with the alloy absorption being essentially a linear combination of the pure Sn and Ti compositions. These compounds are stabilized at high percentages of Sn (x of ∼60\%), as expected, with bromide compositions demonstrating greater stability compared to the iodides. Overall, we find that these materials behave akin to molecular aggregates, with the thermodynamic and optoelectronic properties governed by the intraoctahedral interactions.},
+	urldate = {2023-10-23},
+	journal = {The Journal of Physical Chemistry C},
+	author = {Liga, Shanti M. and Kavanagh, Seán R. and Walsh, Aron and Scanlon, David O. and Konstantatos, Gerasimos},
+	month = oct,
+	year = {2023},
+	note = {Publisher: American Chemical Society},
+	file = {ACS Full Text Snapshot:/Users/kavanase/Zotero/storage/RLFR4VF8/acs.jpcc.html:text/html},
+}
