### 2023-10-24 SMTG-Bham/easyunfold#35 @srmnitc
Documentation does not specify how to run tests
This is part of review for JOSS over at https://github.com/openjournals/joss-reviews/issues/5974

There are automated tests within the repository, which is great! However it is not explained in the documentation how to install the dependencies for the tests, or how to run the tests themselves manually. Adding a small section explaining this could be quite useful.
