### 2023-10-24 SMTG-Bham/easyunfold#36 @srmnitc
How to seek support for the software?
This is part of the JOSS review over at https://github.com/openjournals/joss-reviews/issues/5974

This is a minor point, but the community guidelines section of the review calls for: Contribute to the software, Report issues or problems with the software and Seek support. The first two aspects are already covered in the documentation. Adding a small section about how to reach out to the devs for support would help to close this issue.
