### 2023-10-27 92d17f1
Fix earlier errors in commensurability check (revealed by testing with example data)
