### 2023-10-27 3e219bb
Update `--matrix` help message so users know what form it should be in
