### 2023-11-06 e34cef9
Fix wrapping kpoints at -0.5/0.5 (#39)
Enforce that the kpoints to be wrapped between [-0.5, 0.5), as the
previous code may generated both -0.5 and 0.5 due to finite precision.
This is a achieved by check if the wrapped point is close to 0.5
(<1e-7). If so, it will be changed to -0.5.
