### 2023-11-16 174c32b
Reduce redundant code in tests
