+++ tests/test_cli.py
@@ -20,13 +20,34 @@ def test_generate(si_project_dir):
     tmp_dir = si_project_dir(None)
     os.chdir(tmp_dir)
 
-    output = runner.invoke(easyunfold,
-                           ['generate', 'Si/POSCAR', 'Si_super_deformed/POSCAR', 'KPOINTS_band_low', '--out-file', 'test', '-y'])
+    output = runner.invoke(
+        easyunfold,
+        [
+            'generate',
+            'Si/POSCAR',
+            'Si_super_deformed/POSCAR',
+            'KPOINTS_band_low',
+            '--out-file',
+            'test',
+            '-y',
+        ],
+    )
     assert output.exit_code == 0
 
     output = runner.invoke(
         easyunfold,
-        ['generate', 'Si/POSCAR', 'Si_super_deformed/POSCAR', 'KPOINTS_band_low', '--out-file', 'test', '--nk-per-split', '3', '-y'])
+        [
+            'generate',
+            'Si/POSCAR',
+            'Si_super_deformed/POSCAR',
+            'KPOINTS_band_low',
+            '--out-file',
+            'test',
+            '--nk-per-split',
+            '3',
+            '-y',
+        ],
+    )
     assert output.exit_code == 0
 
     kpts = read_kpoints('KPOINTS_test')[0]
@@ -37,10 +58,20 @@ def test_generate(si_project_dir):
 
     # Test with SCF kpoints
 
-    output = runner.invoke(easyunfold, [
-        'generate', 'Si/POSCAR', 'Si_super_deformed/POSCAR', 'KPOINTS_band_low', '--out-file', 'test', '--scf-kpoints',
-        'Si_super_deformed_soc/IBZKPT', '-y'
-    ])
+    output = runner.invoke(
+        easyunfold,
+        [
+            'generate',
+            'Si/POSCAR',
+            'Si_super_deformed/POSCAR',
+            'KPOINTS_band_low',
+            '--out-file',
+            'test',
+            '--scf-kpoints',
+            'Si_super_deformed_soc/IBZKPT',
+            '-y',
+        ],
+    )
     assert output.exit_code == 0
     kpts, _, _, weights = read_kpoints('KPOINTS_test')
     assert weights[0] == 1.0
@@ -51,7 +82,7 @@ def test_generate(si_project_dir):
 
 def test_generate_agsbte2(agsbte2_project_dir):
     """
-    Test the generate function
+    Test the generate function with SQS AgSbTe2 – in particular the ability to handle slightly-incommensurate cells
     """
     runner = CliRunner()
 
@@ -60,53 +91,75 @@ def test_generate_agsbte2(agsbte2_project_dir):
 
     output = runner.invoke(
         easyunfold,  # ~2.5% lattice mismatch in this case, print warning but continue fine
-        ['generate', 'POSCAR_prim', 'SQS_POSCAR', 'KPOINTS_band', '--out-file', 'test'])
+        ['generate', 'POSCAR_prim', 'SQS_POSCAR', 'KPOINTS_band', '--out-file', 'test'],
+    )
     assert output.exit_code == 0
 
-    assert ('Warning: There is a lattice parameter mismatch in the range 2-5% between the primitive (multiplied by the '
+    _kpts = _check_output_info_and_kpoints_agsbte2(
+        [
+            'Warning: There is a lattice parameter mismatch in the range 2-5% between the primitive (multiplied by the '
             'transformation matrix) and the supercell. This will lead to some quantitative inaccuracies in the '
-            'Brillouin Zone spacing (and thus effective masses) of the unfolded band structures.' in output.output)
-    assert '(Guessed) Transform matrix:\n[[1.0, -0.0, 0.0], [1.0, -3.0, 1.0], [1.0, 1.0, -3.0]]' in output.output
[…]
