### 2023-11-16 4254f5a
Update all SMTG-UCL to SMTG-Bham links
