+++ pyproject.toml
@@ -19,8 +19,8 @@ dependencies = ["scipy~=1.0", "numpy~=1.0", "matplotlib~=3.0", "ase~=3.14", "spg
                 "castepxbin~=0.3.0", "castepinput~=0.1"]
 
 [project.urls]
-"Homepage" = "https://github.com/SMTG-UCL/easyunfold"
-"Documentation" = "https://smtg-ucl.github.io/easyunfold/"
+"Homepage" = "https://github.com/SMTG-Bham/easyunfold"
+"Documentation" = "https://smtg-Bham.github.io/easyunfold/"
 
 [project.scripts]
 easyunfold = "easyunfold.cli:easyunfold"
