+++ .github/workflows/docs-deploy.yaml
@@ -48,7 +48,7 @@ jobs:
       - name: Deploy
         uses: peaceiris/actions-gh-pages@v3
         # Deply the documentation built only on the push even
-        if: ${{ (github.event_name != 'pull_request') && (github.repository == 'SMTG-UCL/easyunfold') }}
+        if: ${{ (github.event_name != 'pull_request') && (github.repository == 'SMTG-Bham/easyunfold') }}
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
           publish_dir: ./docs/_build/html
