+++ _sources/publications.md
@@ -1,19 +1,0 @@
-# Publications
-
-We'll add papers that use `easyunfold` to this list as they come out!
-
-* Nicolson, A.; Breternitz, J.; Kavanagh, S. R.; Tomm, Y.; Morita, K.; Squires, A. G.; Tovar, M.; Walsh,
-  A.; Schorr, S.; Scanlon, D. O. Interplay of Static and Dynamic Disorder in the Mixed-Metal 
-  Chalcohalide Sn$_2$SbS$_2$I$_3$. _J. Am. Chem. Soc._ **2023**. https://doi.org/10.1021/jacs.2c13336.
-
-* Huang, Y.-T.; Kavanagh, S. R.; Righetto, M.; Rusu, M.; Levine, I.; Unold, T.; Zelewski, S. J.; Sneyd, 
-  A. J.; Zhang, K.; Dai, L.; Britton, A. J.; Ye, J.; Julin, J.; Napari, M.; Zhang, Z.; Xiao, J.; 
-  Laitinen, M.; Torrente-Murciano, L.; Stranks, S. D.; Rao, A.; Herz, L. M.; Scanlon, D. O.; Walsh, A.; 
-  Hoye, R. L. Z. Strong Absorption and Ultrafast Localisation in NaBiS$_2$ Nanocrystals with Slow 
-  Charge-Carrier Recombination. _Nat Commun_ **2022**, 13 (1), 4960. 
-* https://doi.org/10.1038/s41467-022-32669-3.
-
-* Wang, Y.; Kavanagh, S. R.; Burgués-Ceballos, I.; Walsh, A.; Scanlon, D. O.; Konstantatos, G. Cation 
-  Disorder Engineering Yields AgBiS$_2$ Nanocrystals with Enhanced Optical Absorption for Efficient 
-  Ultrathin Solar Cells. _Nat. Photon._ **2022**, 16 (3), 235–241.
-  https://doi.org/10.1038/s41566-021-00950-4.
