+++ _sources/apidocs/easyunfold/easyunfold.utils.md
@@ -1,163 +1,0 @@
-# {py:mod}`easyunfold.utils`
-
-```{py:module} easyunfold.utils
-```
-
-```{autodoc2-docstring} easyunfold.utils
-:allowtitles:
-```
-
-## Module Contents
-
-### Functions
-
-````{list-table}
-:class: autosummary longtable
-:align: left
-
-* - {py:obj}`write_kpoints <easyunfold.utils.write_kpoints>`
-  - ```{autodoc2-docstring} easyunfold.utils.write_kpoints
-    :summary:
-    ```
-* - {py:obj}`write_kpoints_castep <easyunfold.utils.write_kpoints_castep>`
-  - ```{autodoc2-docstring} easyunfold.utils.write_kpoints_castep
-    :summary:
-    ```
-* - {py:obj}`write_kpoints_vasp <easyunfold.utils.write_kpoints_vasp>`
-  - ```{autodoc2-docstring} easyunfold.utils.write_kpoints_vasp
-    :summary:
-    ```
-* - {py:obj}`read_kpoints_line_vasp <easyunfold.utils.read_kpoints_line_vasp>`
-  - ```{autodoc2-docstring} easyunfold.utils.read_kpoints_line_vasp
-    :summary:
-    ```
-* - {py:obj}`read_kpoints <easyunfold.utils.read_kpoints>`
-  - ```{autodoc2-docstring} easyunfold.utils.read_kpoints
-    :summary:
-    ```
-* - {py:obj}`read_kpoints_vasp <easyunfold.utils.read_kpoints_vasp>`
-  - ```{autodoc2-docstring} easyunfold.utils.read_kpoints_vasp
-    :summary:
-    ```
-* - {py:obj}`read_kpoints_castep <easyunfold.utils.read_kpoints_castep>`
-  - ```{autodoc2-docstring} easyunfold.utils.read_kpoints_castep
-    :summary:
-    ```
-* - {py:obj}`wrap_kpoints <easyunfold.utils.wrap_kpoints>`
-  - ```{autodoc2-docstring} easyunfold.utils.wrap_kpoints
-    :summary:
-    ```
-* - {py:obj}`find_unique <easyunfold.utils.find_unique>`
-  - ```{autodoc2-docstring} easyunfold.utils.find_unique
-    :summary:
-    ```
-* - {py:obj}`reduce_kpoints <easyunfold.utils.reduce_kpoints>`
-  - ```{autodoc2-docstring} easyunfold.utils.reduce_kpoints
-    :summary:
-    ```
-* - {py:obj}`kpoints_equal <easyunfold.utils.kpoints_equal>`
-  - ```{autodoc2-docstring} easyunfold.utils.kpoints_equal
-    :summary:
-    ```
-````
-
-### Data
-
-````{list-table}
-:class: autosummary longtable
-:align: left
-
-* - {py:obj}`RE_COMMENT <easyunfold.utils.RE_COMMENT>`
-  - ```{autodoc2-docstring} easyunfold.utils.RE_COMMENT
-    :summary:
-    ```
-````
-
-### API
-
-````{py:data} RE_COMMENT
-:canonical: easyunfold.utils.RE_COMMENT
-:value: >
-   None
-
-```{autodoc2-docstring} easyunfold.utils.RE_COMMENT
-```
-
-````
-
-````{py:function} write_kpoints(kpoints: typing.Union[numpy.ndarray, list], outpath, *args, code='vasp', **kwargs)
-:canonical: easyunfold.utils.write_kpoints
-
-```{autodoc2-docstring} easyunfold.utils.write_kpoints
-```
-````
-
[…]
