+++ _sources/examples.md
@@ -1,13 +1,0 @@
-# Examples
-
-Worked examples showcasing the capability of the `easyunfold`.
-Input files for these examples can be found 
-[here](https://github.com/SMTG-UCL/easyunfold/tree/main/examples).
-
-```{toctree}
-:maxdepth: 1
-examples/example_si222.md
-examples/example_mgo.md
-examples/example_nabis2.md
-examples/example_si211_castep.md
-```
