+++ _sources/apidocs/index.rst
@@ -1,11 +1,0 @@
-API Reference
-=============
-
-This page contains auto-generated API reference documentation [#f1]_.
-
-.. toctree::
-   :titlesonly:
-
-   easyunfold/easyunfold
-
-.. [#f1] Created with `sphinx-autodoc2 <https://github.com/chrisjsewell/sphinx-autodoc2>`_

[…]
