### 2023-11-16 6ca6079
Add sphinx autoapi requirement
