### 2023-11-16 7b44ab4
Fix `astroid` requirement (version 3 has known bug)
