+++ pyproject.toml
@@ -28,7 +28,8 @@ easyunfold = "easyunfold.cli:easyunfold"
 [project.optional-dependencies]
 doc = [
     "sphinx-click>=4,<5", "sphinx-autodoc2>=0.4,<0.5", "sphinx-book-theme>=1,<2",
-    "sphinx-copybutton>=0.3,<0.4", "myst-parser[linkify]", "sphinx>=6,<7", "sphinx-autoapi==3.0.0"
+    "sphinx-copybutton>=0.3,<0.4", "myst-parser[linkify]", "sphinx>=6,<7", "astroid<3"
+    # astroid needs to be <3 due to: https://github.com/sphinx-extensions2/sphinx-autodoc2/issues/31
 ]
 test = ["pytest", "pytest-cov", "coverage", "sumo"]
 dos = ["sumo"]
