+++ _sources/apidocs/easyunfold/easyunfold.unfold.md
@@ -1,0 +1,517 @@
+# {py:mod}`easyunfold.unfold`
+
+```{py:module} easyunfold.unfold
+```
+
+```{autodoc2-docstring} easyunfold.unfold
+:allowtitles:
+```
+
+## Module Contents
+
+### Classes
+
+````{list-table}
+:class: autosummary longtable
+:align: left
+
+* - {py:obj}`UnfoldKSet <easyunfold.unfold.UnfoldKSet>`
+  - ```{autodoc2-docstring} easyunfold.unfold.UnfoldKSet
+    :summary:
+    ```
+* - {py:obj}`Unfold <easyunfold.unfold.Unfold>`
+  - ```{autodoc2-docstring} easyunfold.unfold.Unfold
+    :summary:
+    ```
+````
+
+### Functions
+
+````{list-table}
+:class: autosummary longtable
+:align: left
+
+* - {py:obj}`get_symmetry_dataset <easyunfold.unfold.get_symmetry_dataset>`
+  - ```{autodoc2-docstring} easyunfold.unfold.get_symmetry_dataset
+    :summary:
+    ```
+* - {py:obj}`find_K_from_k <easyunfold.unfold.find_K_from_k>`
+  - ```{autodoc2-docstring} easyunfold.unfold.find_K_from_k
+    :summary:
+    ```
+* - {py:obj}`rotate_kpt <easyunfold.unfold.rotate_kpt>`
+  - ```{autodoc2-docstring} easyunfold.unfold.rotate_kpt
+    :summary:
+    ```
+* - {py:obj}`expand_K_by_symmetry <easyunfold.unfold.expand_K_by_symmetry>`
+  - ```{autodoc2-docstring} easyunfold.unfold.expand_K_by_symmetry
+    :summary:
+    ```
+* - {py:obj}`LorentzSmearing <easyunfold.unfold.LorentzSmearing>`
+  - ```{autodoc2-docstring} easyunfold.unfold.LorentzSmearing
+    :summary:
+    ```
+* - {py:obj}`GaussianSmearing <easyunfold.unfold.GaussianSmearing>`
+  - ```{autodoc2-docstring} easyunfold.unfold.GaussianSmearing
+    :summary:
+    ```
+* - {py:obj}`make_kpath <easyunfold.unfold.make_kpath>`
+  - ```{autodoc2-docstring} easyunfold.unfold.make_kpath
+    :summary:
+    ```
+* - {py:obj}`clean_latex_string <easyunfold.unfold.clean_latex_string>`
+  - ```{autodoc2-docstring} easyunfold.unfold.clean_latex_string
+    :summary:
+    ```
+* - {py:obj}`spectral_function_from_weight_sets <easyunfold.unfold.spectral_function_from_weight_sets>`
+  - ```{autodoc2-docstring} easyunfold.unfold.spectral_function_from_weight_sets
+    :summary:
+    ```
+* - {py:obj}`spectral_weight_multiple_source <easyunfold.unfold.spectral_weight_multiple_source>`
+  - ```{autodoc2-docstring} easyunfold.unfold.spectral_weight_multiple_source
+    :summary:
+    ```
+* - {py:obj}`concatenate_scf_kpoints <easyunfold.unfold.concatenate_scf_kpoints>`
+  - ```{autodoc2-docstring} easyunfold.unfold.concatenate_scf_kpoints
+    :summary:
+    ```
+* - {py:obj}`create_white_colormap <easyunfold.unfold.create_white_colormap>`
+  - ```{autodoc2-docstring} easyunfold.unfold.create_white_colormap
+    :summary:
+    ```
+* - {py:obj}`create_white_colormap_from_existing <easyunfold.unfold.create_white_colormap_from_existing>`
[…]
