+++ _sources/apidocs/easyunfold/easyunfold.cli.md
@@ -1,0 +1,224 @@
+# {py:mod}`easyunfold.cli`
+
+```{py:module} easyunfold.cli
+```
+
+```{autodoc2-docstring} easyunfold.cli
+:allowtitles:
+```
+
+## Module Contents
+
+### Functions
+
+````{list-table}
+:class: autosummary longtable
+:align: left
+
+* - {py:obj}`easyunfold <easyunfold.cli.easyunfold>`
+  - ```{autodoc2-docstring} easyunfold.cli.easyunfold
+    :summary:
+    ```
+* - {py:obj}`generate <easyunfold.cli.generate>`
+  - ```{autodoc2-docstring} easyunfold.cli.generate
+    :summary:
+    ```
+* - {py:obj}`unfold <easyunfold.cli.unfold>`
+  - ```{autodoc2-docstring} easyunfold.cli.unfold
+    :summary:
+    ```
+* - {py:obj}`unfold_status <easyunfold.cli.unfold_status>`
+  - ```{autodoc2-docstring} easyunfold.cli.unfold_status
+    :summary:
+    ```
+* - {py:obj}`unfold_calculate <easyunfold.cli.unfold_calculate>`
+  - ```{autodoc2-docstring} easyunfold.cli.unfold_calculate
+    :summary:
+    ```
+* - {py:obj}`add_mpl_style_option <easyunfold.cli.add_mpl_style_option>`
+  - ```{autodoc2-docstring} easyunfold.cli.add_mpl_style_option
+    :summary:
+    ```
+* - {py:obj}`unfold_effective_mass <easyunfold.cli.unfold_effective_mass>`
+  - ```{autodoc2-docstring} easyunfold.cli.unfold_effective_mass
+    :summary:
+    ```
+* - {py:obj}`add_plot_options <easyunfold.cli.add_plot_options>`
+  - ```{autodoc2-docstring} easyunfold.cli.add_plot_options
+    :summary:
+    ```
+* - {py:obj}`unfold_plot <easyunfold.cli.unfold_plot>`
+  - ```{autodoc2-docstring} easyunfold.cli.unfold_plot
+    :summary:
+    ```
+* - {py:obj}`process_dos <easyunfold.cli.process_dos>`
+  - ```{autodoc2-docstring} easyunfold.cli.process_dos
+    :summary:
+    ```
+* - {py:obj}`unfold_plot_projections <easyunfold.cli.unfold_plot_projections>`
+  - ```{autodoc2-docstring} easyunfold.cli.unfold_plot_projections
+    :summary:
+    ```
+* - {py:obj}`_unfold_plot <easyunfold.cli._unfold_plot>`
+  - ```{autodoc2-docstring} easyunfold.cli._unfold_plot
+    :summary:
+    ```
+* - {py:obj}`print_symmetry_data <easyunfold.cli.print_symmetry_data>`
+  - ```{autodoc2-docstring} easyunfold.cli.print_symmetry_data
+    :summary:
+    ```
+* - {py:obj}`matrix_from_string <easyunfold.cli.matrix_from_string>`
+  - ```{autodoc2-docstring} easyunfold.cli.matrix_from_string
+    :summary:
+    ```
+````
+
+### Data
+
+````{list-table}
+:class: autosummary longtable
+:align: left
+
+* - {py:obj}`SUPPORTED_DFT_CODES <easyunfold.cli.SUPPORTED_DFT_CODES>`
+  - ```{autodoc2-docstring} easyunfold.cli.SUPPORTED_DFT_CODES
+    :summary:
+    ```
+* - {py:obj}`DEFAULT_CMAPS <easyunfold.cli.DEFAULT_CMAPS>`
+  - ```{autodoc2-docstring} easyunfold.cli.DEFAULT_CMAPS
+    :summary:
+    ```
[…]
