+++ _sources/apidocs/easyunfold/easyunfold.md
@@ -1,0 +1,51 @@
+# {py:mod}`easyunfold`
+
+```{py:module} easyunfold
+```
+
+```{autodoc2-docstring} easyunfold
+:allowtitles:
+```
+
+## Submodules
+
+```{toctree}
+:titlesonly:
+:maxdepth: 1
+
+easyunfold.wavefun
+easyunfold.utils
+easyunfold.vasp_constant
+easyunfold.effective_mass
+easyunfold.wavecar
+easyunfold.cli
+easyunfold.plotting
+easyunfold.procar
+easyunfold.unfold
+```
+
+## Package Contents
+
+### Data
+
+````{list-table}
+:class: autosummary longtable
+:align: left
+
+* - {py:obj}`__version__ <easyunfold.__version__>`
+  - ```{autodoc2-docstring} easyunfold.__version__
+    :summary:
+    ```
+````
+
+### API
+
+````{py:data} __version__
+:canonical: easyunfold.__version__
+:value: >
+   '0.3.0'
+
+```{autodoc2-docstring} easyunfold.__version__
+```
+
+````

[…]
