### 2023-11-22 SMTG-Bham/easyunfold#40 @srmnitc
Suggestions for the publication
This is part of the JOSS review over at https://github.com/openjournals/joss-reviews/issues/5974

- Line 10-11, 14 could be strengthened by a citation.
- Line 36, k-points were not mentioned before, the theory comes only later. Probably a one line introduction could be really useful.
- Line 47, VASP and CASTEP needs citations.
