### 2023-11-27 SMTG-Bham/easyunfold#41 @awvwgk
Compatibility with other DFT packages
Is easyunfold compatible with other DFT packages than Vasp and Castep. The documentation states that integration is supposed to be straightforward.

An example via an ASE calculator which saves the band structure would be a very convincing example for a widely compatible band structure unfolding implementation. Alternatively, a short note in the docs and/or software paper why the ASE provided band structure tools are insufficient for easyunfold would be useful.
