### 2023-11-27 #5974 @awvwgk
I finished my review as well, easyunfold provides a comprehensive documentation with examples, tutorials, theory, and technical reference. The repository provides all the necessary parts to install the package and to follow the documentation. Besides the already raised issues, I opened a request to evaluate the need for tailored implementation of IO interfaces to Vasp and Castep, especially with a standard package like ASE which can handle and extract band structures already (see https://github.com/SMTG-Bham/easyunfold/issues/41). Otherwise this package is good to publish in JOSS.
