### 2023-12-03 3ffbda0
Docs and paper update for JOSS
