+++ paper/paper.bib
@@ -272,3 +272,61 @@ publisher: Nature Publishing Group},
 	note = {Publisher: American Chemical Society},
 	file = {ACS Full Text Snapshot:/Users/kavanase/Zotero/storage/RLFR4VF8/acs.jpcc.html:text/html},
 }
+
+@article{Ganose2018, doi = {10.21105/joss.00717}, url = {https://doi.org/10.21105/joss.00717}, year = {2018}, publisher = {The Open Journal}, volume = {3}, number = {28}, pages = {717}, author = {Alex M. Ganose and Adam J. Jackson and David O. Scanlon}, title = {sumo: Command-line tools for plotting and analysis of periodic *ab initio* calculations}, journal = {Journal of Open Source Software} }
+
+@article{Kim_2020,
+doi = {10.1088/2515-7655/aba081},
+url = {https://dx.doi.org/10.1088/2515-7655/aba081},
+year = {2020},
+month = {jul},
+publisher = {IOP Publishing},
+volume = {2},
+number = {3},
+pages = {036001},
+author = {Sunghyun Kim and Samantha N Hood and Ji-Sang Park and Lucy D Whalley and Aron Walsh},
+title = {Quick-start guide for first-principles modelling of point defects in crystalline materials},
+journal = {Journal of Physics: Energy},
+abstract = {Defects influence the properties and functionality of all crystalline materials. For instance, point defects participate in electronic (e.g. carrier generation and recombination) and optical (e.g. absorption and emission) processes critical to solar energy conversion. Solid-state diffusion, mediated by the transport of charged defects, is used for electrochemical energy storage. First-principles calculations of defects based on density functional theory have been widely used to complement, and even validate, experimental observations. In this ‘quick-start guide’, we discuss the best practice in how to calculate the formation energy of point defects in crystalline materials and analysis techniques appropriate to probe changes in structure and properties relevant across energy technologies.}
+}
+
+@article{vasp1,
+  title = {Efficient iterative schemes for ab initio total-energy calculations using a plane-wave basis set},
+  author = {Kresse, G. and Furthm\"uller, J.},
+  journal = {Phys. Rev. B},
+  volume = {54},
+  issue = {16},
+  pages = {11169--11186},
+  numpages = {0},
+  year = {1996},
+  month = {Oct},
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRevB.54.11169},
+  url = {https://link.aps.org/doi/10.1103/PhysRevB.54.11169}
+}
+@article{vasp2,
+title = {Efficiency of ab-initio total energy calculations for metals and semiconductors using a plane-wave basis set},
+journal = {Computational Materials Science},
+volume = {6},
+number = {1},
+pages = {15-50},
+year = {1996},
+issn = {0927-0256},
+doi = {https://doi.org/10.1016/0927-0256(96)00008-0},
+url = {https://www.sciencedirect.com/science/article/pii/0927025696000080},
+author = {G. Kresse and J. Furthmüller},
[…]
