### 2023-12-03 7097e27
Added explanation regard using ASE's io framework
