### 2023-12-03 1ad6376
Updated URL SMTG-UCL -> SMTG-Bham
