+++ paper/paper.bib
@@ -224,7 +224,7 @@ publisher: Nature Publishing Group},
   year = {2022},
   publisher = {GitHub},
   journal = {GitHub repository},
-  url = {https://github.com/SMTG-UCL/doped}
+  url = {https://github.com/SMTG-Bham/doped}
 }
 
 @article{icet,
