### 2023-12-03 SMTG-Bham/easyunfold#40 @zhubonan
Thanks for the comments. I have updated the paper accordingly.

> Line 10-11, 14 could be strengthened by a citation.

I did not find a suitable citation for describing how to run band structure calculation and its wide use - in most cases, the procedures are described in the documentation of each DFT codes. Nevertheless, I included a citation of a widely used helper tool (at least, in my group..), [sumo](https://smtg-bham.github.io/sumo/index.html), to run band structure calculations and plot the results.
