### 2023-12-03 SMTG-Bham/easyunfold#35 @zhubonan
Thanks, instructions have been added to the docs, along with some notes about development:

https://github.com/SMTG-Bham/easyunfold/blob/1ae2b7546e74d9656fee82ae20a759c5ad84d64c/docs/index.md?plain=1#L123C10-L140
