### 2023-12-03 SMTG-Bham/easyunfold#40 @zhubonan
> Line 36, k-points were not mentioned before, the theory comes only later. Probably a one line introduction could be really useful.

Good point, I think it is not necessary to introduce the concept of k-points here. The relevant sentence has been updated to:

https://github.com/SMTG-Bham/easyunfold/blob/1ae2b7546e74d9656fee82ae20a759c5ad84d64c/paper/paper.md?plain=1#L61-L62
