### 2023-12-04 ca15570
Update bib file
