+++ paper/paper.bib
@@ -312,7 +312,7 @@ number = {1},
 pages = {15-50},
 year = {1996},
 issn = {0927-0256},
-doi = {https://doi.org/10.1016/0927-0256(96)00008-0},
+doi = {10.1016/0927-0256(96)00008-0},
 url = {https://www.sciencedirect.com/science/article/pii/0927025696000080},
 author = {G. Kresse and J. Furthmüller},
 abstract = {We present a detailed description and comparison of algorithms for performing ab-initio quantum-mechanical calculations using pseudopotentials and a plane-wave basis set. We will discuss: (a) partial occupancies within the framework of the linear tetrahedron method and the finite temperature density-functional theory, (b) iterative methods for the diagonalization of the Kohn-Sham Hamiltonian and a discussion of an efficient iterative method based on the ideas of Pulay's residual minimization, which is close to an order Natoms2 scaling even for relatively large systems, (c) efficient Broyden-like and Pulay-like mixing methods for the charge density including a new special ‘preconditioning’ optimized for a plane-wave basis set, (d) conjugate gradient methods for minimizing the electronic free energy with respect to all degrees of freedom simultaneously. We have implemented these algorithms within a powerful package called VAMP (Vienna ab-initio molecular-dynamics package). The program and the techniques have been used successfully for a large number of different systems (liquid and amorphous semiconductors, liquid simple and transition metals, metallic and semi-conducting surfaces, phonons in simple metals, transition metals and semiconductors) and turned out to be very reliable.}
@@ -326,7 +326,7 @@ pages = {567--570},
 volume = {220},
 number = {5-6},
 journal = {Zeitschrift für Kristallographie - Crystalline Materials},
-doi = {doi:10.1524/zkri.220.5.567.65075},
+doi = {10.1524/zkri.220.5.567.65075},
 year = {2005},
 lastchecked = {2023-12-03}
 }
