+++ paper/paper.bib
@@ -273,7 +273,7 @@ publisher: Nature Publishing Group},
 	file = {ACS Full Text Snapshot:/Users/kavanase/Zotero/storage/RLFR4VF8/acs.jpcc.html:text/html},
 }
 
-@article{Ganose2018, doi = {10.21105/joss.00717}, url = {https://doi.org/10.21105/joss.00717}, year = {2018}, publisher = {The Open Journal}, volume = {3}, number = {28}, pages = {717}, author = {Alex M. Ganose and Adam J. Jackson and David O. Scanlon}, title = {sumo: Command-line tools for plotting and analysis of periodic *ab initio* calculations}, journal = {Journal of Open Source Software} }
+@article{Ganose_2018, doi = {10.21105/joss.00717}, url = {https://doi.org/10.21105/joss.00717}, year = {2018}, publisher = {The Open Journal}, volume = {3}, number = {28}, pages = {717}, author = {Alex M. Ganose and Adam J. Jackson and David O. Scanlon}, title = {sumo: Command-line tools for plotting and analysis of periodic *ab initio* calculations}, journal = {Journal of Open Source Software} }
 
 @article{Kim_2020,
 doi = {10.1088/2515-7655/aba081},
