### 2023-12-11 93900ee
minor edits on paper.md
