### 2023-12-17 23bab9f
Merge Paper branch (#43)
Merge content from the paper branch into the main branch. Minor updates of the documentation, the actual code is unchanged.


Co-authored-by: Seán Kavanagh <51478689+kavanase@users.noreply.github.com>
Co-authored-by: Sean Kavanagh <s.kavanagh19@imperial.ac.uk>
Co-authored-by: Mojtaba Barzegari <40744245+mbarzegary@users.noreply.github.com>
