+++ paper/paper.md
@@ -1,0 +1,152 @@
+---
+title: 'easyunfold: A Python package for unfolding electronic band structures'
+tags:
+  - Python
+  - materials science
+  - first-principles
+  - density functional theory
+authors:
+  - name: Bonan Zhu
+    orcid: 0000-0001-5601-6130
+    equal-contrib: false
+    affiliation: "1, 2, 3" # (Multiple affiliations must be quoted)
+  - name: Seán R. Kavanagh
+    orcid: 0000-0003-4577-9647
+    equal-contrib: false
+    affiliation: "1, 2, 4" # (Multiple affiliations must be quoted)
+  - name: David Scanlon
+    orcid: 0000-0001-9174-8601
+    equal-contrib: false
+    affiliation: "1, 2, 3" # (Multiple affiliations must be quoted)
+affiliations:
+ - name: Department of Chemistry, University College London, London, United Kingdom
+   index: 1
+ - name: Thomas Young Centre, University College London, London, United Kingdom
+   index: 2
+ - name: The Faraday Institution, Didcot, United Kingdom
+   index: 3
+ - name: Department of Materials, Imperial College London, London, United Kingdom
+   index: 4
+date: 6 June 2023
+bibliography: paper.bib
+---
+
+# Summary
+
+The electronic band structure is an important property for
+understanding and designing solid crystalline materials in many fields
+such as photovoltaic, catalytic, thermoelectric and transparent-conducting
+materials. Obtaining the band structure for an ideal crystal through first-principles
+density functional theory (DFT) calculations is a well-established routine operation[@Ganose_2018].
+However, the materials of interest are often complex and the simulation cells may contain multiple primitive
+cells of the archetypal structure when, for example, modelling disordered or defective materials[@Kim_2020].
+Repeating the unit cell in real space results in folded band structures, as illustrated in \autoref{fig:figure1}, complicating its interpretation and analysis.
+Band structure unfolding maps the electronic structure from supercell calculations back to the reciprocal lattice of the primitive cell,
+thereby enabling researchers to understand structure-property relationships and compare the effect of various crystal imperfections on an equal footing.
+With `easyunfold`, we provide a simple, easy-to-use, yet powerful and flexible tool which implements the band structure unfolding workflow using plane-wave DFT codes, from input file generation to publication-quality plotting.
+
+![Schematics of band folding in a 2D lattice: (a) The primitive cell in real space; (b) The $2\times2$ supercell in real space. (c) A slice ($\varepsilon$) of the band structure in the Brillouin zone of the primitive cell; (d) The same band structure slice ($\varepsilon$) in the Brillouin zone of the supercell. Band structure unfolding is the process of transforming the band structure slice in (d) back to the Brillouin zone of the primitive cell (i.e. the form shown in (c)). \label{fig:figure1}](figure1.png){width=100mm}
+
+# Statement of need
+
+There are existing packages that provide band structure unfolding capabilities, such as `BandUp`
+[@bandup], and `VaspBandUnfolding` [@vaspbandunfolding].
+`easyunfold` is written in Python with a focus on user-friendliness, data provenance, reproducibility, and publication-quality figure generation.
+An example output of the effective band structure produced is shown in \autoref{fig:figure2} for a $2\times2\times2$ $\mathrm{MgO}$ supercell containing a neutral oxygen vacancy.
[…]
