+++ paper/compile.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+docker run --rm \
+    --volume $PWD:/data \
+    --user $(id -u):$(id -g) \
+    --env JOURNAL=joss \
+    openjournals/inara
