### 2023-12-17 #5974 @zhubonan
> - [x]  Make a tagged release of your software, and list the version tag of the archived version here.

Sure, I have made a release, the release version tag is `v0.3.4` and the version number of the actual code is `0.3.4`.

> - [x] Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
> - [x] Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
>  - [x] Please list the DOI of the archived version here.

The archived DOI is [10.5281/zenodo.10396925](https://doi.org/10.5281/zenodo.10396925)
