### 2024-01-03 #5974 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
