### 2024-01-11 b3e0852
Improve the effective mass detection algorithm
Make the algorithm more robust - now filter by intensity first.
Also added ability to manually pass the extrema points when
computing effective masses.
