+++ easyunfold/cli.py
@@ -267,19 +267,17 @@ def add_mpl_style_option(func):
 @click.option('--spin', type=int, default=0, help='Index of the spin channel.', show_default=True)
 @click.option('--npoints', type=int, default=3, help='Number of kpoints used for fitting from the extrema.', show_default=True)
 @click.option('--extrema-detect-tol', type=float, default=0.01, help='Tolerance for band extrema detection.', show_default=True)
-@click.option('--degeneracy-detect-tol',
-              type=float,
-              default=0.01,
-              help='Tolerance for band degeneracy detection at extrema.',
-              show_default=True)
 @click.option('--nocc', type=int, help='DEV: Use this band as the extrema at all kpoints.')
 @click.option('--plot', is_flag=True, default=False)
 @click.option('--plot-fit', is_flag=True, default=False, help='Generate plots of the band edge and parabolic fits.')
 @click.option('--fit-label', help='Which branch to use for plot fitting. e.g. electrons:0', default='electrons:0', show_default=True)
 @click.option('--band-filter', default=None, type=int, help='Only displace information for this band.')
 @click.option('--out-file', '-o', default='unfold-effective-mass.png', help='Name of the output file.', show_default=True)
-def unfold_effective_mass(ctx, intensity_threshold, spin, band_filter, npoints, extrema_detect_tol, degeneracy_detect_tol, nocc, plot,
-                          plot_fit, fit_label, out_file):
+@click.option('--emin', type=float, default=-5., help='Minimum energy in eV relative to the reference.', show_default=True)
+@click.option('--emax', type=float, default=5., help='Maximum energy in eV relative to the reference.', show_default=True)
+@click.option('--manual-extrema', help='Manually specify the extrema to use for fitting, in the form "mode,k_index,band_index"')
+def unfold_effective_mass(ctx, intensity_threshold, spin, band_filter, npoints, extrema_detect_tol, nocc, plot, plot_fit, fit_label,
+                          out_file, emin, emax, manual_extrema):
     """
     Compute and print effective masses by tracing the unfolded weights.
 
@@ -293,7 +291,7 @@ def unfold_effective_mass(ctx, intensity_threshold, spin, band_filter, npoints, 
     from easyunfold.unfold import UnfoldKSet
     from tabulate import tabulate
     unfoldset: UnfoldKSet = ctx.obj['obj']
-    efm = EffectiveMass(unfoldset, intensity_tol=intensity_threshold, extrema_tol=extrema_detect_tol, degeneracy_tol=degeneracy_detect_tol)
+    efm = EffectiveMass(unfoldset, intensity_tol=intensity_threshold, extrema_tol=extrema_detect_tol)
 
     click.echo('Band extrema data:')
     table = []
@@ -306,12 +304,20 @@ def unfold_effective_mass(ctx, intensity_threshold, spin, band_filter, npoints, 
 
     if nocc:
         efm.set_nocc(nocc)
-    output = efm.get_effective_masses(ispin=spin, npoints=npoints)
+    if manual_extrema is None:
+        output = efm.get_effective_masses(ispin=spin, npoints=npoints)
+    else:
+        mode, ik, ib = manual_extrema.split(',')
+        ik = int(ik)
+        ib = int(ib)
+        click.echo(f'Using manually passed kpoint and band: {ik},{ib}')
+        output = efm.get_effective_masses(ispin=spin, npoints=npoints, mode=mode, iks=[ik], iband=[[ib]])
 
     # Filter by band if requested
     if band_filter is not None:
         for carrier in ['electrons', 'holes']:
-            output[carrier] = [entry for entry in output[carrier] if entry['band_index'] == band_filter]
[…]

[…]
