### 2024-01-11 ef7d197
Added effective mass example to MgO
