### 2024-01-11 dea07fb
Fix typos and small update to notebook
