### 2024-01-11 #5974 @Kevin-Mattheus-Moerman
@zhubonan as AEiC I will now help to process this work for acceptance in JOSS. I have checked your repository, this review, the archive link, and the paper. Most seems in order. However, I do have the below points that require your attention.

- [ ] Please modify the Zenodo listed version tag to read: `v0.3.4`, i.e. include the `v`.
- [ ] You use `analyse` and `colour`, Brittish English, but also `serialization`, American English, consider using one consistently.
