### 2024-01-12 4e38de7
Update docstrings, docs and CLI outputs
