+++ easyunfold/effective_mass.py
@@ -89,10 +89,11 @@ class EffectiveMass:
         """
         Instantiate the object
 
-        Args:
-            unfold (UnfoldKSet): The ``UnfoldKSet`` object that holes unfolding data.
-            intensity_tol (float): Intensity tolerance for detecting band edges.
-            parabolic (bool): Perform parabolic fit or not. Defaults to True as non-parabolic fit is not working at the moment...
+        :param unfold: The ``UnfoldKSet`` object that holds unfolding data.
+        :param intensity_tol: Intensity threshold for detecting band edges.
+        :param extrema_tol: Distance tolerance for detecting band edges.
+        :param parabolic: Perform parabolic fit or not. The default is None.
+        :param npoints: The number of points used for fitting.
         """
         self.unfold: UnfoldKSet = unfold
         self.intensity_tol = intensity_tol
@@ -109,26 +110,36 @@ class EffectiveMass:
 
     @property
     def kpoints(self):
+        """
+        The primitive cell k-points used for unfolding.
+        """
         return self.unfold.kpts_pc
 
     @property
     def kpoints_labels(self):
+        """
+        The primitive cell k-points labels set for unfolding.
+        """
         return self.unfold.kpoint_labels
 
     def get_band_extrema(self, mode: str = 'cbm', extrema_tol: float = None, ispin=0):
         """
-        Obtain the kpoint idx of band extrema, sub indices in th set and the band indices.
+        Obtain the kpoint idx of band extrema, sub indices in the set, and the band indices.
 
-        The search takes two steps, first the kpoints at the band extrema is located by comparing the
-        band energies with that recorded in supplied *cbm* and *vbm*, based on the `extrema_tol`.
-        Afterwards, the band indices are selected at the these kpoints using
+        The search takes two steps. First, the kpoints at the band extrema are located by comparing the
+        band energies with that recorded in the supplied *cbm* and *vbm*, based on the `extrema_tol`.
+        Afterwards, the band indices are selected at these kpoints using the `tol` set.
 
-        Returns:
-            A tuple of extrema locations including a list of kpoint indices, sub-indices within
-            the set and the band indices at each kpoint that is within the `tol` set.
+        :param mode: The mode to search for band extrema. Can be either 'cbm' (conduction band minimum) or 'vbm'
+                     (valence band maximum).
+        :param extrema_tol: The tolerance for determining the proximity of band energies to the cbm/vbm.
+                            If not provided, the default tolerance from `self.extrema_detect_tol` is used.
+        :param ispin: The spin index. Default is 0.
+        :returns: A tuple of extrema locations including a list of kpoint indices, sub-indices within
+                 the set, and the band indices at each kpoint that is within the `extrema_tol` from the cbm/vbm.
+        :raises ValueError: If an unknown mode is provided.
         """
-        if extrema_tol is None:
-            extrema_tol = self.extrema_detect_tol
+        extrema_tol = self.extrema_detect_tol if extrema_tol is None else extrema_tol
 
         intensity_tol = self.intensity_tol
 
@@ -179,6 +190,13 @@ class EffectiveMass:
     def _get_fitting_data(self, kidx: int, iband: int, direction=1, ispin=0, npoints=None):
         """
         Get fitting data for a specific combination of kpoint and band index
+
+        :param kidx: The index of the kpoint
+        :param iband: The index of the band
+        :param direction: The direction of the data collection, defaults to 1
+        :param ispin: The index of the spin, defaults to 0
+        :param npoints: Override for the number of data points to collect
+        :returns: The normalized kpoint distances, normalized effective energies, and the original kpoint distances and effective energies
         """
         istart = kidx
[…]
