+++ tests/test_effective_mass.py
@@ -3,6 +3,7 @@ Test the effective mass functionality
 """
 from pathlib import Path
 import pytest
+import numpy as np
 from monty.serialization import loadfn
 import easyunfold.effective_mass as em
 
@@ -42,16 +43,15 @@ def test_effective_mass(effective_mass_obj):
     assert len(kdist) == len(effective_mass_obj.kpoints)
 
     fdata = effective_mass_obj._get_fitting_data(0, 16, 1, 0, 3)
-    assert (fdata[0] == kdist[:3]).all()
-    assert len(fdata[1]) == 3
+    assert len(fdata[1]) == 6
 
     assert len(output['electrons']) == 2
     assert len(output['holes']) == 2
 
     assert output['electrons'][0]['kpoint_label_from'] == '\\Gamma'
     assert output['electrons'][0]['kpoint_label_to'] == 'L'
-    assert output['electrons'][0]['effective_mass'] == pytest.approx(0.39912256690278236)
+    assert output['electrons'][0]['effective_mass'] == pytest.approx(0.36959097872)
 
     assert output['holes'][0]['kpoint_label_from'] == '\\Gamma'
     assert output['holes'][0]['kpoint_label_to'] == 'L'
-    assert output['holes'][0]['effective_mass'] == pytest.approx(-5.972424721183893)
+    assert output['holes'][0]['effective_mass'] == pytest.approx(-3.36071124861)
