### 2024-01-12 908b189
Implement separate folder mode for split k-points
