### 2024-01-14 88aafb3
Update cell matrix convention notes
