### 2024-01-14 bc98ec5
Updated spelling and citations
