### 2024-01-14 918a729
effective mass fitting improve (#44)
* Improve the effective mass detection algorithm

Make the algorithm more robust - now filter by intensity first.
Also added ability to manually pass the extrema points when
computing effective masses.

* Added effective mass example to MgO

* Added ipynb example for effective mass

* Added link to the notebook example.

* Fix pylint complaint

* Fix typos and small update to notebook

* Update docstrings, docs and CLI outputs

* Fix tests

* Implement separate folder mode for split k-points

* Update cell matrix convention notes

---------

Co-authored-by:
[…]
