### 2024-01-14 b12c5dc
Update joss paper files (#45)
* Bump version number

* Fix typos

* Updated spelling and citations

---------

Co-authored-by: Seán Kavanagh <51478689+kavanase@users.noreply.github.com>
