+++ paper/paper.bib
@@ -90,7 +90,6 @@ Vegge and Lasse Vilhelmsen and Michael Walter and Zhenhua Zeng and Karsten W Jac
 	year = {2012},
 	note = {Publisher: American Physical Society},
 	pages = {085201},
-	file = {Popescu_Zunger_2012_Extracting \$E\$ versus \$-stackrel P-vec k \$ effective band structure from.pdf:C\:\\Users\\zhubo\\Dropbox (UCL)\\Zotero\\Physical Review B\\Popescu_Zunger_2012_Extracting \$E\$ versus \$-stackrel P-vec k \$ effective band structure from.pdf:application/pdf},
 }
 
 
@@ -131,7 +130,6 @@ Vegge and Lasse Vilhelmsen and Michael Walter and Zhenhua Zeng and Karsten W Jac
 	note = {Number: 1
 Publisher: Nature Publishing Group},
 	keywords = {easyunfold, Electronic materials, Solar cells},
-	file = {Huang et al_2022_Strong absorption and ultrafast localisation in NaBiS2 nanocrystals with slow.pdf:C\:\\Users\\zhubo\\Dropbox (UCL)\\Zotero\\Nature Communications2022\\Huang et al_2022_Strong absorption and ultrafast localisation in NaBiS2 nanocrystals with slow.pdf:application/pdf},
 }
 
 @article{nicolson:2023,
@@ -146,7 +144,6 @@ Publisher: Nature Publishing Group},
 	date = {2023-05-30},
 	note = {Publisher: American Chemical Society},
 	keywords = {easyunfold},
-	file = {ACS Full Text Snapshot:C\:\\Users\\zhubo\\Zotero\\storage\\XAU4UB2M\\jacs.html:text/html;Nicolson et al_2023_Interplay of Static and Dynamic Disorder in the Mixed-Metal Chalcohalide.pdf:C\:\\Users\\zhubo\\Dropbox (UCL)\\Zotero\\Journal of the American Chemical Society2023\\Nicolson et al_2023_Interplay of Static and Dynamic Disorder in the Mixed-Metal Chalcohalide.pdf:application/pdf},
 }
 
 @article{wang:2022,
@@ -167,24 +164,27 @@ Publisher: Nature Publishing Group},
 	note = {Number: 3
 Publisher: Nature Publishing Group},
 	keywords = {easyunfold, Quantum dots, Solar cells},
-	file = {Wang et al_2022_Cation disorder engineering yields AgBiS2 nanocrystals with enhanced optical.pdf:C\:\\Users\\zhubo\\Dropbox (UCL)\\Zotero\\Nature Photonics2022\\Wang et al_2022_Cation disorder engineering yields AgBiS2 nanocrystals with enhanced optical.pdf:application/pdf},
 }
 
 
 @article{huber:2020,
 	title = {{AiiDA} 1.0, a scalable computational infrastructure for automated reproducible workflows and data provenance},
-	url = {http://arxiv.org/abs/2003.12476},
-	journaltitle = {{arXiv}:2003.12476 [cond-mat]},
+	volume = {7},
+	copyright = {2020 The Author(s)},
+	issn = {2052-4463},
+	url = {https://www.nature.com/articles/s41597-020-00638-4},
+	doi = {10.1038/s41597-020-00638-4},
[…]
