### 2024-01-15 #5974 @zhubonan
> @zhubonan as an author (who happens to also be an editor) please refrain from using the editor commands here.

Err sorry...

> please edit the archive listed license to match your software license.

Sure, I have just updated the archive's licence to MIT to match the code.
