### 2024-01-15 #5974 @Kevin-Mattheus-Moerman
@openjournals/dev can we make it that an author cannot run the editor command for the issue they are listed as author for (such that the commands do not work even if the submitting author happens to be a JOSS editor?)?
