### 2024-01-19 #5974 @arfon
> @openjournals/dev can we make it that an author cannot run the editor command for the issue they are listed as author for (such that the commands do not work even if the submitting author happens to be a JOSS editor?)?

I understand the request, but I think this is best handled with training/guidelines rather than code.
